+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-results-seo-tool/dot-results-seo-tool.component.spec.ts
@@ -1,59 +1,7 @@
 import { Spectator, byTestId, createComponentFactory } from '@ngneat/spectator';
 
 import { DotResultsSeoToolComponent } from './dot-results-seo-tool.component';
-
-const seoOGTagsMock = {
-    description:
-        'Get down to Costa Rica this winter for some of the best surfing int he world. Large winter swell is pushing across the Pacific.',
-    language: 'english',
-    favicon: 'http://localhost:8080/application/themes/landing-page/img/favicon.ico',
-    title: 'A title',
-    author: 'dotCMS',
-    copyright: 'dotCMS LLC, Miami Florida, US',
-    'og:title': 'A title',
-    'og:url': 'https://dotcms.com$!{dotPageContent.canonicalUrl}',
-    'og:image': 'https://dotcms.com/images/default.png'
-};
-
-const seoOGTagsResultMock = [
-    {
-        key: 'Favicon',
-        keyIcon: 'pi-check-circle',
-        keyColor: 'var(--color-alert-green)',
-        items: [
-            { message: 'FavIcon found!', color: 'var(--color-alert-green)', itemIcon: 'pi-check' }
-        ],
-        sort: 1
-    },
-    {
-        key: 'Description',
-        keyIcon: 'pi-exclamation-triangle',
-        keyColor: 'var(--color-alert-red)',
-        items: [
-            {
-                message: 'Meta Description not found! Showing Description instead.',
-                color: 'var(--color-alert-red)',
-                itemIcon: 'pi-times'
-            }
-        ],
-        sort: 2,
-        info: "The length of the description allowed will depend on the reader's device size; on the smallest size only about 110 characters are allowed."
-    },
-    {
-        key: 'Title',
-        keyIcon: 'pi-exclamation-triangle',
-        keyColor: 'var(--color-alert-yellow)',
-        items: [
-            {
-                message: 'HTML Title found, but has fewer than 30 characters of content.',
-                color: 'var(--color-alert-yellow)',
-                itemIcon: 'pi-exclamation-circle'
-            }
-        ],
-        sort: 3,
-        info: 'HTML Title content should be between 30 and 60 characters.'
-    }
-];
+import { seoOGTagsMock, seoOGTagsResultMock } from './mocks';
 
 describe('DotResultsSeoToolComponent', () => {
     let spectator: Spectator<DotResultsSeoToolComponent>;
