+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-results-seo-tool/dot-results-seo-tool.component.stories.ts
@@ -1,0 +1,37 @@
+import { Meta, Story, moduleMetadata } from '@storybook/angular';
+
+import { NgClass, NgFor, NgIf, TitleCasePipe } from '@angular/common';
+
+import { CardModule } from 'primeng/card';
+
+import { DotMessageService } from '@dotcms/data-access';
+
+import { DotResultsSeoToolComponent } from './dot-results-seo-tool.component';
+import { seoOGTagsMock, seoOGTagsResultMock } from './mocks';
+
+export default {
+    title: 'DotCMS/Results Seo Tool',
+    component: DotResultsSeoToolComponent,
+    args: {
+        hostName: 'A title',
+        seoOGTags: seoOGTagsMock,
+        seoOGTagsResults: seoOGTagsResultMock
+    },
+    decorators: [
+        moduleMetadata({
+            imports: [NgClass, CardModule, NgFor, TitleCasePipe, NgIf],
+            providers: [
+                {
+                    provide: DotMessageService,
+                    useValue: {
+                        get: () => 'Copy'
+                    }
+                }
+            ]
+        })
+    ]
+} as Meta;
+
+export const Primary: Story = (args) => ({
+    props: args
+});
